+++ src/test/java/org/scify/jedai/configuration/StepByStepGridConfigurationCCER.java
@@ -48,8 +48,8 @@ import org.scify.jedai.utilities.datastructures.BilateralDuplicatePropagation;
 public class StepByStepGridConfigurationCCER {
 
     static float getTotalComparisons(List<AbstractBlock> blocks) {
-        int originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
+        long originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Long::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
